Never show the intro screens
The hack we did at the last minute actually introduced another issue: After the introduction is completed, the last step is just a black screen… and the user actually needs to click a top right X button to close, which is not very visible and won’t actually work in mobile devices.

We didn’t catch this issue earlier because the original implementation was based on localStorage information and would only be shown at the first time the game was open. Since we were always with the game open… we were never seeing it.

Today while showing the game to friends the issue became very apparent. 

And while it is minor… it hides all the hard work we did behind this black screen.